#1658 Added unit test.
